Version 0.5.0

+++ utilpaisagem/scenery/tile.py
@@ -1,3 +1,4 @@
+from typing import Union
 from numbers import Number
 from pathlib import Path
 from urllib.error import URLError, ContentTooShortError
@@ -5,11 +6,15 @@ from queue import Queue
 from PIL import Image
 from threading import Thread
 from decimal import Decimal
+from datetime import datetime, timedelta
 import math, tempfile, shutil, os, configparser, ast
 from babel.numbers import format_decimal
 from utilpaisagem.scenery.common import Coordinates, DOWNLOAD_RES, MIN_RES, MAX_RES
 from utilpaisagem.scenery.image_service import ImageService
-from utilpaisagem.gui.common import format_status, Settings, QUIT
+from utilpaisagem.gui.common import format_status, Settings, TileColors, QUIT
+
+class NeedsRenewalError(Exception):
+    pass
 
 class Tile(object):
     """
@@ -21,7 +26,8 @@ class Tile(object):
     index:int
     coordinates:Coordinates
     resolution:int
-    upstream_queue:Queue|None
+    upstream_queue:Union[Queue|None]
+    tile_manager_queue:Union[Queue|None]
     settings:Settings
 
     def __init__(
@@ -30,7 +36,7 @@ class Tile(object):
         lat:Number=float('nan'),
         lon:Number=float('nan'), 
         resolution:int=DOWNLOAD_RES,
-        upstream_queue:Queue|None=None,
+        upstream_queue:Union[Queue|None]=None,
     ):
         """
         New Tile object defined by either index or by coordinates.
@@ -167,7 +173,26 @@ class Tile(object):
             Path(f'{lon_dir}{abs(math.floor(self.coordinates.lon_left)):03}' + \
             f'{lat_dir}{abs(math.floor(self.coordinates.lat_bottom)):02}')
 
-    def retrieve(self, path:Path, image_service:ImageService, download_res=DOWNLOAD_RES, compress='smart', upstream_queue:Queue=None):
+    def is_failed(self, log:configparser.ConfigParser, logpath:Path):
+        if log['INFO']['success'] != 'True' or ast.literal_eval(log['INFO']['failures']) != []:
+            return True
+        return False
+
+    def is_old(self, log:configparser.ConfigParser, logpath:Path):
+        if datetime.now() - datetime.fromtimestamp(os.path.getmtime(logpath)) > \
+            timedelta(days=self.settings.renewal_age):
+            return True
+        return False
+
+    def retrieve(
+        self,
+        path:Path,
+        image_service:ImageService,
+        download_res:int=DOWNLOAD_RES,
+        compress:str='smart',
+        upstream_queue:Union[Queue|None]=None,
+        tile_manager_queue:Union[Queue|None]=None
+    ):
         """
         Tests if the image exists and is not needed to regenerate it. If Ok, touch the
         file in order to know that it has been used. The image should be generated again if it is smaller than the demanded
@@ -178,6 +203,8 @@ class Tile(object):
             image_service(ImagerService): image downloader
             download_res(int): exponent of two representing vertical image size
             compress(str): compression method, either 'png', 'dds' or 'smart'. Defaults to 'smart'
+            upstream_queue(Queue): queue for user communication
+            tile_manager_queue(Queue): queue for the TileManager
         """
         if not image_service.can_download(self.coordinates):
             if self.upstream_queue is None:
@@ -262,6 +289,7 @@ class Tile(object):
         if logpath.exists():
             log = configparser.ConfigParser()
             log.read(logpath)
+            delete = False
             try:
                 if int(log['INFO']['resolution']) < self.resolution:
                     if self.upstream_queue is None:
@@ -272,7 +300,7 @@ class Tile(object):
                             self
                         ))
                     raise AssertionError
-                if log['INFO']['success'] != 'True' or ast.literal_eval(log['INFO']['failures']) != []:
+                if self.is_failed(log, logpath):
                     if self.upstream_queue is None:
                         print('Failure on previous download. Downloading again.')
                     else:
@@ -305,6 +333,8 @@ class Tile(object):
                             self    
                         ))
                     raise AssertionError
+                if self.is_old(log, logpath):
+                    raise NeedsRenewalError
                 if self.upstream_queue is None:
                     print(f'Tile {self.index} has already been downloaded. Skipping.')
                 else:
@@ -324,6 +354,17 @@ class Tile(object):
                             _('Failed to check previous download ("{e}"). Downloading again.').format(e=e),
                             self
                         ))
+                delete = True
+            except NeedsRenewalError:
+                if self.upstream_queue is None:
+                    print(f'Tile {self.index} needs renewal. Downloading again.')
+                else:
+                    self.upstream_queue.put_nowait(format_status(
+                        _('Tile {index} needs renewal. Downloading again.').format(index=self.index),
+                        self,
+                    ))
+                delete = True
+            if delete:
                 # Remove previously downloaded file
                 if Path(path / f'{self.index}.png').exists():
                     Path(path / f'{self.index}.png').unlink()
@@ -405,8 +446,28 @@ class Tile(object):
                 }
                 with open(path / f'{self.index}.log', 'w') as logfile:
                     log.write(logfile)
+                try:
+                    tile_manager_queue.put_nowait((
+                        self.index,
+                        self.coordinates,
+                        self.get_path(self.settings.orthophotos_folder) / filename.name,
+                        TileColors.good,
+                        str(self.get_path(Path())).split('/'),
+                    ))
+                except AttributeError:
+                    pass
 
         except (URLError, ContentTooShortError) as e:
+            try:
+                tile_manager_queue.put_nowait((
+                    self.index,
+                    self.coordinates,
+                    self.get_path(self.settings.orthophotos_folder) / filename.name,
+                    TileColors.failed,
+                    str(self.get_path(Path())).split('/'),
+                ))
+            except AttributeError:
+                pass
             if self.resolution > MIN_RES:
                 if self.upstream_queue is None:
                     print(f'Error downloading cell {self.index}[{line}][{cell}]: {e}.')
@@ -440,6 +501,18 @@ class Tile(object):
             if QUIT.is_set:
                 return
             else:
+                if (self.get_path(path) / (self.index + '.png')).exists() or \
+                    (self.get_path(path) / (self.index + '.dds')).exists():
+                    try:
+                        tile_manager_queue.put_nowait((
+                            self.index,
+                            self.coordinates,
+                            self.get_path(self.settings.orthophotos_folder) / filename.name,
+                            TileColors.failed,
+                            str(self.get_path(Path())).split('/'),
+                        ))
+                    except AttributeError:
+                        pass
                 raise e
 
     def files_exist(self) -> bool:
@@ -456,16 +529,20 @@ class Tile(object):
         return False
 
 
-    def delete_files(self):
+    def delete_files(self, tile_manager_queue:Queue):
+        size = 0
         dds = self.get_path(self.settings.orthophotos_folder).glob(f'{self.index}.dds', case_sensitive=False)
         for file in dds:
+            size += os.path.getsize(file)
             file.unlink()
         png = self.get_path(self.settings.orthophotos_folder).glob(f'{self.index}.png', case_sensitive=False)
         for file in png:
+            size += os.path.getsize(file)
             file.unlink()
         log = self.get_path(self.settings.orthophotos_folder).glob(f'{self.index}.log', case_sensitive=False)
         for file in log:
             file.unlink()
+        tile_manager_queue.put_nowait((-self.index, self.coordinates, size, str(self.get_path(Path())).split('/')))
 
     @classmethod
     def tile_width(cls, lat):
